Refactorizar folder constants
