--- src/constants/QUERY_KEY_PRODUCTS.js
@@ -1,1 +1,0 @@
-export const QUERY_KEY_PRODUCTS = "products";
