+++ src/constants/index.js
@@ -1,3 +1,2 @@
-export * from "./API_URL";
-export * from "./QUERY_KEY_CATEGORIES";
-export * from "./QUERY_KEY_PRODUCTS";
+export * from "./api";
+export * from "./querieskey";
