+++ src/constants/querieskey.js
@@ -1,1 +1,2 @@
 export const QUERY_KEY_CATEGORIES = "categories";
+export const QUERY_KEY_PRODUCTS = "products";
